@@ -16,22 +16,15 @@ public class CustomViewPager extends ViewPager {
 
     @Override
     public boolean onTouchEvent(MotionEvent event) {
-        if (this.isPagingEnabled) {
-            return super.onTouchEvent(event);
-        }
-
-        return false;
+        return this.isPagingEnabled && super.onTouchEvent(event);
     }
 
     @Override
     public boolean onInterceptTouchEvent(MotionEvent event) {
-        if (this.isPagingEnabled) {
-            return super.onInterceptTouchEvent(event);
-        }
-
-        return false;
+        return this.isPagingEnabled && super.onInterceptTouchEvent(event);
     }
 
+    @SuppressWarnings("unused")
     public void setPagingEnabled(boolean b) {
         this.isPagingEnabled = b;
     }
